- move DetailedBranchPage.js and GeneralBranchPage.js to Page folder
